refactor: rename loader to mixin
