+++ lib/base_loader.js
@@ -238,12 +238,12 @@ class EggLoader {
 const loaders = [
   require('./plugin_loader'),
   require('./config_loader'),
-  require('./loader/extend'),
-  require('./loader/custom'),
+  require('./mixin/extend'),
+  require('./mixin/custom'),
   require('./proxy_loader'),
   require('./service_loader'),
-  require('./loader/middleware'),
-  require('./loader/controller'),
+  require('./mixin/middleware'),
+  require('./mixin/controller'),
 ];
 
 for (const loader of loaders) {
