+++ lib/mixin/extend.js
